admins can no longer delete themselves and fixed a bug

+++ Project/RefugEAP-Project/src/main/java/com/example/refugeapproject/membership/UserRepo.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.repository.query.Param;
 
-public interface UserRepo extends CrudRepository<User,Integer> {
+public interface UserRepo extends CrudRepository<User,Long> {
 
     @Query("SELECT u FROM User u WHERE u.username = :username")
     public User getUserByUsername(@Param("username") String username);
